install/0000_90_cluster-version-operator_02_servicemonitor: Add alert informing that minor release 4.8 entered maintenance phase
Add an alert `MinorReleaseEnteredMaintenanceSupport` informing the user
that the current cluster minor release 4.8 has entered its maintenance
support phase. This will notify the users so they are aware of this
situation, and they can plan for upgrading to a newer minor release
before the end of the maintenance phase.

Set the severity to `info` as this situation doesn't require immediate
action, but the user should be aware of this information.

Provide a link to the documentation containing more information
regarding the life cycle phases. Provide the link in the description
since the field `runbook_url` is intended to contain a link that guides
a cluster owner through the steps of fixing a problem. And this link
only provides additional information.